Changed the float inputs so they allow point for decimals instead of comas
Also, fixed a couple bugs with those float inputs

+++ TFG/GUI Control/PerceptionGUI.cs
@@ -15,7 +15,7 @@ public class PerceptionGUI : GUIElement
     /// <summary>
     /// Parameter for the Timer Perception
     /// </summary>
-    public int timerNumber;
+    public float timerNumber;
 
     /// <summary>
     /// Name of the selected <see cref="FSM"/> or <see cref="BehaviourTree"/> for IsInState and BehaviourTreeStatus Perceptions
